% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data_2PL_1.R
\docType{data}
\name{example_data_2PL_1}
\alias{example_data_2PL_1}
\title{Binary Response Dataset for Federated 2PL Model}
\format{
A dataframe with 81 rows and 10 columns.
Each row corresponds to a student's set of responses, and each column represents the responses of all students to a particular item.
}
\source{
The data are synthetically generated.
}
\usage{
example_data_2PL_1
}
\description{
A synthetic dataset composed of responses from 81 students to 10 items, with all responses binary.
It is the first part of "example_data_2PL".
It is used to test the correctness of federated 2PL model.
A response of 1 indicates a correct answer, and 0 represents an incorrect answer.
}
\examples{
data(example_data_2PL_1)

}
\keyword{datasets}
