% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data_graded_and_binary.R
\docType{data}
\name{example_data_graded_and_binary}
\alias{example_data_graded_and_binary}
\title{Graded Response Dataset for Federated Graded Model}
\format{
A dataframe with 160 rows and 8 columns. Each row denotes a student's responding status, and each column denotes the responding status for all students in one item.
}
\source{
The data were generated synthetically for illustrative purposes and not representing any real answering status.
}
\usage{
example_data_graded_and_binary
}
\description{
A synthetic dataset containing 160 students and 8 items, with the first item graded response (0-3 points) and other items binary response.
It could be used in the federated graded model testing.
}
\examples{
data(example_data_graded_and_binary)

}
\keyword{datasets}
