% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedirt_1PL_data.R
\name{fedirt_1PL_data}
\alias{fedirt_1PL_data}
\title{Federated 1PL model}
\usage{
fedirt_1PL_data(inputdata)
}
\arguments{
\item{inputdata}{A List of all responding matrices.}
}
\value{
A list with the estimated global difficulty b, person's abilities ability, sites' abilities site, and log-likelihood value loglik.
}
\description{
This function is used to test the accuracy and processing time of this algorithm. It inputs a list of responding matrices and return the federated 1PL parameters.
}
\details{
Input is a List of responding matrices from each school, every responding matrix is one site's data.
}
\examples{
inputdata = list(as.matrix(example_data_2PL))
fedresult = fedirt_1PL_data(inputdata)

inputdata = list(as.matrix(example_data_2PL_1), as.matrix(example_data_2PL_2))
fedresult = fedirt_1PL_data(inputdata)

}
