% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedirt_2PL_schooleffects.R
\name{fedirt_2PL_schooleffects}
\alias{fedirt_2PL_schooleffects}
\title{Federated 2PL model with school effects}
\usage{
fedirt_2PL_schooleffects(inputdata)
}
\arguments{
\item{inputdata}{A List of all responding matrices.}
}
\value{
A list with the estimated global discrimination a, global difficulty b, person's abilities ability, sites' abilities site, and log-likelihood value loglik. It also displays the school ability sc, which is considered as a fixed effect.
}
\description{
This function is used to test the accuracy and processing time of this algorithm. It inputs a list of responding matrices and return the federated 2PL parameters.
Note: To use federated 2PL in distributed datasets, please use fedirt_2PL().
}
\details{
Input is a List of responding matrices from each school, every responding matrix is one site's data.
}
\examples{
inputdata = list(as.matrix(example_data_2PL))
fedresult = fedirt_2PL_schooleffects(inputdata)

inputdata = list(as.matrix(example_data_2PL_1), as.matrix(example_data_2PL_2))
fedresult = fedirt_2PL_schooleffects(inputdata)

}
