% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personfit.R
\name{personfit}
\alias{personfit}
\title{Federated IRT person fit}
\usage{
personfit(fedresult)
}
\arguments{
\item{fedresult}{fedirt result object}
}
\value{
a list of person fit in each school.
}
\description{
personfit calculates the Zh values, infit and outfit statistics.
The returned object is a list.
}
\details{
Input is the object of fedirt class.
}
\examples{
# turn input data to a list
inputdata = list(as.matrix(example_data_2PL))
# Call fedirt() function, and use 2PL model
fedresult = fedirt(inputdata, model_name = "2PL")
personfitResult = personfit(fedresult)
}
