% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personscore.R
\name{personscore}
\alias{personscore}
\title{Federated IRT person score}
\usage{
personscore(fedresult)
}
\arguments{
\item{fedresult}{fedirt result object}
}
\value{
a list of person score in each school.
}
\description{
This function calculates persons' ability.
}
\details{
Input is the object of fedirt class.
}
\examples{
# turn input data to a list
inputdata = list(as.matrix(example_data_2PL))
# Call fedirt() function, and use 2PL model
fedresult = fedirt(inputdata, model_name = "2PL")
personscoreResult = personscore(fedresult)
}
