% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{SE}
\alias{SE}
\title{Federated IRT SE}
\usage{
SE(fedresult)
}
\arguments{
\item{fedresult}{fedirt result object}
}
\value{
An array of standard errors for all parameters.
}
\description{
Calculates Standard Error(SE) for FedIRT models.
}
\details{
Input is the object of fedirt class.
}
\examples{
# turn input data to a list
inputdata = list(as.matrix(example_data_2PL))
# Call fedirt() function, and use 2PL model
fedresult = fedirt(inputdata, model_name = "2PL")
# get SE result
SEresult = SE(fedresult)
}
