% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedirt.R
\name{summary.fedirt}
\alias{summary.fedirt}
\title{Summary Method for FedIRT Objects}
\usage{
\method{summary}{fedirt}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fedirt}.}

\item{...}{Additional arguments passed to the summary method.}
}
\description{
Provides a summary for objects of class \code{fedirt}.
}
